Rename estimate-gas file to get-gas-params
